gui_file_server/requirements.txt
tkinter
Pillow>=9.0.0
humanize>=4.0.0
pathlib2